fixed 64 bits warning
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@1441 689a6050-402a-0410-94f2-e92a70836424

+++ parser/parse_privacy.c
@@ -210,7 +210,7 @@ int parse_privacy(struct sip_msg *msg)
 	return -1;
     }
 
-    msg->privacy->parsed = (void *)values;
+    msg->privacy->parsed = (void *)(long)values;
 
     return 0;
